requirements.txt
pyTelegramBotAPI==4.22.0
google-generativeai==0.8.4
requests==2.31.0